refactor(ui): extract shared inline-create logic for the bind pickers
Address review feedback on PR #2938:

- Extract the create-bridge orchestration (displayedAgents bridge, empty
  seed, justCreated/creating state) into a shared useInlineAgentCreate
  hook, and the picker row into a shared BindAgentRow — removing the
  near-verbatim Slack/Telegram copy-paste before it can drift.
- Default template picks the first non-experimental entry explicitly,
  rather than depending on the catalogue's sort order.
- Note the deliberate controlled-state (not RHF+Zod) deviation in
  CreateAgentInline.
